% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisProb.R
\name{analysisProb}
\alias{analysisProb}
\title{Calculate the Probability of an Observed p-value for a miRNA Family for a subset of RNAseq Genes}
\usage{
analysisProb(mockAnalysisTable, analysisTable, miRNAFamily)
}
\arguments{
\item{analysisTable}{The output of the miRsiftAnalyze function.}

\item{miRNAFamily}{The 7 nucleotide seed that defines a given miRNA family.}

\item{simAnalysisTable}{The output of the miRsiftSimAnalyze function.}
}
\description{
This function can be used following miRsiftSimAnalyze in order to determine the probability that a p-value for a particular miRNA family would be observed for a randomly selected subset of RNAseq genes. Useful for determining if including PT regulation enhances the targeting signature seen for a particular miRNA family. Specifically, this function calculates the proportion of p-values for a given miRNA family with a p-value at least as significant as that observed for the miRNA family in miRsift analysis on the subset of RNAseq genes showing evidence of PT regulation.
}
\examples{
analysisProb(mockAnalysisTable, analysisTable, "GAGGUAG") #determines the probability that the p-value for let-7 in miRsift analysis on the subset of PT regulated genes would also be observed for a random subset of genes of the same size.
}
\keyword{miRsift}
\keyword{probability}

