% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCPMCutoff.R
\name{calcCPMCutoff}
\alias{calcCPMCutoff}
\title{Calculates CPM Cutoff}
\usage{
calcCPMCutoff(libSizes, minDepth)
}
\arguments{
\item{minDepth}{The minimum read depth that the user wants to use.}

\item{listObject}{Vector of library sizes for sequencing data set.}
}
\description{
Calculates CPM cutoff for seq data given a read depth cutoff. Also requires a vector of library 
sizes. Specifically, first determines the library size minimum, then calculates the CPM value
that corresponds to the minimum read depth for the smallest-sized library.
}
\examples{
cutoff <- calcCPMCutoff(c(1000000,2000000), 6)
}
\keyword{CPMcutoff}

