% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importPT.R
\name{importPT}
\alias{importPT}
\title{Import Optional List of Genes with Post-Transcriptional Regulatory Signatures for miRsift Analysis}
\usage{
importPT(PTFile)
}
\arguments{
\item{PTFile}{This file is provided by the user. It should contain no header. Each line of the file should contain a single gene symbol, corresponding to a gene that was determined to post-transcriptionally regulated in the given dataset and comparison as determined by an analysis such as EISA.}
}
\description{
This function imports genes with post-transcriptional regulatory signatures from a file supplied by the user. This list is optional but recommended for miRsift analysis.
}
\examples{
PTList <- importPT(PTFile)
}
\keyword{EISA}
\keyword{posttranscriptional}

