% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSeq.R
\name{importSeq}
\alias{importSeq}
\title{Import Sequencing Data for miRsift Analysis}
\usage{
importSeq(seqFile, group, contrast, minDepth = 6)
}
\arguments{
\item{seqFile}{This file is provided by the user. It should be in tab-delim format and contain a header line. The first column
should be the name of each gene in gene symbol format. All other columns should be raw, unnormalized
counts in integer format. See miRsift user manual for an example.}

\item{group}{A vector which designates which sample group each of the count containing columns of
the input file corresponds to. See miRsift user manual for an example.}

\item{contrast}{A vector of two values designating which two sample groups to compare. See miRsift 
user manual for an example.}

\item{minDepth}{Optional. A numeric specifying the minimum number of counts required for a gene to be considered
expressed and thus used in our later miRsift analysis. The default value is 6.}
}
\description{
This function reads in sequencing data from a file provided by the user, calculating the log2
fold change values and determining differential expression of expressed genes with edgeR for 
a given comparison.
}
\examples{
rnaseqTable <- importSeq("rnaseq_counts.txt", c(1,1,2,2,3,3), c(1,2)) #analysis comparing sample groups 1 and 2, using default minDepth for filtering
rnaseqTable <- importSeq("rnaseq_counts.txt", c(1,1,2,2,3,3), c(1,3), 10) #analysis comparing sample groups 1 and 3, using custom minDepth of 10 for filtering
}
\references{
Robinson MD, McCarthy DJ and Smyth GK (2010). “edgeR: a Bioconductor package for differential expression analysis of digital gene expression data.” Bioinformatics, 26, pp. -1.
McCarthy, J. D, Chen, Yunshun, Smyth and K. G (2012). “Differential expression analysis of multifactor RNA-Seq experiments with respect to biological variation.” Nucleic Acids Research, 40(10), pp. -9.
Robinson MD and Smyth GK (2007). “Moderated statistical tests for assessing differences in tag abundance.” Bioinformatics, 23, pp. -6.
Robinson MD and Smyth GK (2008). “Small-sample estimation of negative binomial dispersion, with applications to SAGE data.” Biostatistics, 9, pp. -11.
Zhou X, Lindsay H and Robinson MD (2014). “Robustly detecting differential expression in RNA sequencing data using observation weights.” Nucleic Acids Research, 42, pp. e91.
}
\keyword{importseq}

