% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRsiftAnalyze.R
\name{miRsiftAnalyze}
\alias{miRsiftAnalyze}
\title{Perform miRsift Analysis}
\usage{
miRsiftAnalyze(contextTable, rnaseqTable, ...)
}
\arguments{
\item{contextTable}{The output of the miRsift buildContextTable function.}

\item{rnaseqTable}{The output of the miRsift importSeq function for RNAseq data.}

\item{...}{Optional parameters for filtering out lowly expressed miRNAs. These are PTList (output of importPTList), smallrnaseqMinExp (integer), and smallrnaseqTable (output of importSeq function for smallRNAseq data). Both are required in order to filter out lowly expressed miRNAs from miRsift analysis.}
}
\description{
This function performs miRsift analysis. This analysis can be done on only RNAseq data. It can include an optional list of genes identified to be post-transcriptionally (PT) regulated, which will be used to filter the RNAseq data and enhance miRNA regulatory signatures (highly recommended). Finally, small RNAseq data can be included, with the option to use this data in order to filter out lowly expressed miRNA families from the analysis.
}
\examples{
simAnalysisTable <- miRsiftAnalyze(contextTable, rnaseqTable) #miRsift analysis, with all expressed genes from the RNAseq data considered.
simAnalysisTable <- miRsiftAnalyze(contextTable, rnaseqTable, PTList=PTList) #miRsift analysis, with only genes showing PT regulation from the RNAseq data considered.
simAnalysisTable <- miRsiftAnalyze(contextTable, rnaseqTable, PTList=PTList, smallrnaseqTable=smallrnaseqTable) #miRsift analysis, with only genes showing PT regulation from the RNAseq data considered. Correlates results with changes in small RNA sequencing data.
simAnalysisTable <- miRsiftAnalyze(contextTable, rnaseqTable, PTList=PTList, smallrnaseqMinExp=100, smallrnaseqTable=smallrnaseqTable) #miRsift analysis, with only genes showing PT regulation from the RNAseq data considered. Correlates results with changes in small RNA sequencing data, and also refines analysis to only focus on highyl expressed miRNA famiies.
}
\keyword{miRsift}

