% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRsiftSimAnalyze.R
\name{miRsiftSimAnalyze}
\alias{miRsiftSimAnalyze}
\title{Perform a Simulated miRsift Analysis on Random Data Subsets}
\usage{
miRsiftSimAnalyze(contextTable, rnaseqTable, subsetSize, reps, ...)
}
\arguments{
\item{contextTable}{The output of the miRsift buildContextTable function.}

\item{subsetSize}{The number (integer) of genes to be randomly selected for miRsift analysis from all expressed genes. This will be the number of post-transcriptionally regulated genes used for miRsift analysis.}

\item{reps}{The number of times randomly sampling and miRsift analysis is performed in order to create a distribution. 10,000 is the suggested value.}

\item{...}{Optional parameters for filtering out lowly expressed miRNAs. Only include if filtering was performed for miRsift analysis. These are smallrnaseqMinExp (integer) and smallrnaseqTable (output of importSeq function for smallRNAseq data). Both are required in order to filter out lowly expressed miRNAs from miRsift analysis.}

\item{group}{The output of the miRsift importSeq function for RNAseq data.}
}
\description{
This function performs a simulated miRsift analysis on random subsets of expressed genes that can then be compared to miRsift analysis on the subset of genes showing post-transcriptional regulation. This allows the user to determine if using the subset of post-transcriptionally regulated genes is enhancing the signal of their miRsift analysis.
}
\examples{
simAnalysisTable <- miRsiftSimAnalyze(contextTable, rnaseqTable, 250, 10000) #simulated miRsift analysis on randomly sampled subset of 250 genes, performed 10,000 times. All miRNA families are considered.
simAnalysisTable <- miRsiftSimAnalyze(contextTable, rnaseqTable, 250, 10000, smallrnaseqMinExp=100, smallrnaseqTable=smallrnaseqTable) #simulated miRsift analysis on randomly sampled subset of 250 genes, performed 10,000 times. Only miRNA families with at least 100 reads are considered.
}
\keyword{miRsift}
\keyword{simulation}

