% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAnalysis.R
\name{plotAnalysis}
\alias{plotAnalysis}
\title{Visualize Results of miRsift Analysis}
\usage{
plotAnalysis(analysisTable)
}
\arguments{
\item{analysisTable}{The output of the miRsiftAnalyze function.}
}
\description{
This function creates a plot to help visualize the resuls of miRsift analysis.
}
\examples{
plotAnalysis(analysisTable)
}
\keyword{plotmiRsift}

