% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSimAnalysis.R
\name{plotSimAnalysis}
\alias{plotSimAnalysis}
\title{Visualize Results of Simulated miRsift Analysis}
\usage{
plotSimAnalysis(simAnalysisTable)
}
\arguments{
\item{simAnalysisTable}{The output of the miRsiftSimAnalyze function.}
}
\description{
This function creates a plot to help visualize the resuls of simulated miRsift analyses.
}
\examples{
plotAnalysis(simAnalysisTable)
}
\keyword{plotmiRsift}

